pyproject.toml
# PIP, using PEP621
[project]
name = "imc"
authors = [
    {name = "Andre Rendeiro", email = "afrendeiro@gmail.com"},
]
description = "A framework for IMC data analysis."
readme = "README.md"
keywords = [
    "computational biology",
    "bioinformatics",
    "imaging mass cytometry",
    "imaging",
    "mass cytometry",
    "mass spectrometry",
]
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Development Status :: 3 - Alpha",
    "Typing :: Typed",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
licence = "gpt3"
requires-python = ">=3.8"
# not yet supported by pip!
dependencies = [
    "setuptools_scm",
    "ordered-set",
    "urlpath",
    "PyYAML",
    "imctools>=2.1.0",
    "joblib",
    "leidenalg",
    "python-louvain",
    "networkx",
    "pandas>=1.0.1",
    "matplotlib",
    "seaborn",
    "fastcluster",
    "parmap",
    "scanpy",
    "numpy_groupies",
    "tifffile==2020.6.3"
]
dynamic = ['version']

[project.optional-dependencies]
# not yet supported by pip!
deepcell = [
    "DeepCell>=0.8.3,<1.0.0",
]
cellpose = [
    "cellpose>=0.1.0.1,<1.0.0",
]
dev = [
    "wheel",
    "ipython",
    "black[d]",
    "mypy",  # pin to version supporting pyproject.toml
    "pandas-stubs",
    "pylint",
    "git-lint",
    "pytest",
    "pydocstyle",
    "rich",
    # data-science-types
    "PyQt5",
]
test = [
    "pytest>=6",
    "pytest-cov",
]
doc = [
    "Sphinx",
    "sphinx-issues",
    "sphinx-rtd-theme",
    "sphinx-argparse",
]

[project.urls]
homepage = "https://github.com/ElementoLab/imc"
# documentation = "readthedocs.org"
repository = "https://github.com/ElementoLab/imc"
changelog = "https://github.com/ElementoLab/imc/blob/master/docs/source/changelog.md"

[project.entry-points."imc"]
imc = "imc.cli:main"


[build-system]
# requires = ["poetry>=0.12", "setuptools>=45", "wheel", "poetry-dynamic-versioning"]
# build-backend = "poetry.masonry.api"
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]  # 
write_to = "imc/_version.py"
# write_to_template = '__version__ = "{version}"'

# Poetry
[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"

[tool.poetry]
name = "imc"
version = "0.0.0"  # waiting on next release of poetry to use dynamic-versioning extension
description = "A package for the analysis of imaging mass cytometry (IMC) data"
authors = ["Andre Rendeiro <afrendeiro@gmail.com>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
ordered-set = "^4.0.2"
PyYAML = "^5.4.1"
pandas = ">=1.0.1"
tifffile = "^2020.6.3"
imctools = "^2.1.0"
imagecodecs = "^2020.5.30"
colorama = "^0.4.3"
# divvy = "^0.5.0"
h5py = "^2.10.0"
# ordered-set = "^4.0.2"
anndata = "^0.7.3"
scanpy = "^1.5.1"
leidenalg = "^0.8.1"
python-louvain = "^0.14"
networkx = "^2.4"
parmap = "^1.5.2"
joblib = "^0.15.1"

[tool.poetry.dev-dependencies]
ipython = "^7.16.1"
pylint = "^2.5.3"
git-lint = "^0.1.2"
black = {extras = ["d"], version = "^19.10b0"}
mypy = "^0.782"
pytest = "^5.4.3"
Sphinx = "^3.1.1"
sphinx-issues = "^1.2.0"
sphinx-rtd-theme = "^0.5.0"

[tool.poetry.extras]
deepcell = [
    "DeepCell",
]
cellpose = [
    "cellpose",
]

[tool.black]
line-length = 90
target-version = ['py37']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.mypy]
python_version = '3.8'
warn_return_any = true
warn_unused_configs = true

# Packages without type annotations in shed yet
[[tool.mypy.overrides]]
module = [
    'numpy.*',
    'pandas.*',
    'scipy.*',
    'skimage.*',
    'matplotlib.*',
    'seaborn.*',
    'parmap.*',
    'anndata.*',
    'scanpy.*',
    'pymde.*',
    'umap.*',
    'networkx.*',
    'pingouin.*',
    'tqdm.*',
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
    "integration",
]
markers = [
    'slow', # 'marks tests as slow (deselect with "-m 'not slow'")',
    'serial'
]
